Re-factored
